connect dot net to django for user model

+++ backend-api/CampusHostels.API/Application/DTOs/UnitCreateDto.cs
@@ -6,7 +6,7 @@ public class UnitCreateDto
     public int Floor { get; set; }
     public string? RoomNumber { get; set; }
     public string? ImageUrl { get; set; }
-    public int? Cost { get; set; }
+    public decimal? Cost { get; set; }
     public int? MaxNoOfPeople { get; set; }
     public string UnitType { get; set; } = "Single";
 }
