+++ backend-api/CampusHostels.API/Domain/Entities/PaymentSummary.cs
@@ -7,7 +7,7 @@ public class PaymentSummary
     public int Id { get; set; }
     public int TenancyAgreementId { get; set; }
     public TenancyAgreement? TenancyAgreement { get; set; }
-    public int TotalAmountPaid { get; set; }
-    public int AmountLeft { get; set; }
+    public decimal TotalAmountPaid { get; set; }
+    public decimal AmountLeft { get; set; }
     public DateTime? LastPaymentDate { get; set; }
 }
